refactor: migra Stockfish para o modulo acchess, documenta IPC e atualiza testes
