+++ software/acchess/include/fen_generator/fen_generator.hpp
@@ -3,7 +3,6 @@
 
 #include <array>
 #include <string>
-#include "piece_detector/piece_detector.hpp"
 
 namespace ac
 {
