+++ software/acchess/tests/test_stockfish_uci.cpp
@@ -1,25 +1,17 @@
 #include <gtest/gtest.h>
-#include "stockfish_uci.hpp" 
+#include "engine/stockfish_uci.hpp" 
 
 TEST(StockfishUCITest, Inicializacao) {
-    
-    ac::StockfishUCI stockfish("stockfish"); 
-    
-    
+    ac::chess::engine::StockfishUCI stockfish("stockfish");
     bool iniciou = stockfish.init();
-    
     EXPECT_TRUE(iniciou); 
 }
 
 TEST(StockfishUCITest, RetornaMelhorJogada) {
-    
-    ac::StockfishUCI stockfish("stockfish");
-    
-    
+    ac::chess::engine::StockfishUCI stockfish("stockfish");
     stockfish.init();
     
     std::string fen_inicial = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
-    
     std::string melhor_jogada = stockfish.get_best_move(fen_inicial, 5);
     
     ASSERT_FALSE(melhor_jogada.empty());
@@ -27,13 +19,12 @@ TEST(StockfishUCITest, RetornaMelhorJogada) {
 }
 
 TEST(StockfishUCITest, BestMoveIsCheckMate){
-    ac::StockfishUCI stockfish("stockfish");
+    ac::chess::engine::StockfishUCI stockfish("stockfish");
     stockfish.init();
 
     std::string fen = "6k1/5ppp/8/8/8/5Q2/6PP/6K1 w - - 0 1";
-
     std::string best_move = stockfish.get_best_move(fen,15);
 
-    std::cout << "best_move: " + best_move ;
+    std::cout << "best_move: " << best_move << std::endl;
     EXPECT_EQ(best_move, "f3a8");
 }
